Separate feeds methods from messages methods
